Refactor: Move request extraction logic to dedicated module and fix referrer validation
- Created extractFromRequest.js for all request data extraction functions
- Moved extractReferrer, extractToken, getIp and other extraction functions from auth-utils.js
- Added /api/validate-referrer endpoint in auth.pollinations.ai for database referrer checks
- Updated all services to use the new modules
- Fixed circular dependencies between modules
- Updated auth.pollinations.ai to correctly handle referrer validation
