+++ shared/auth-utils.js
@@ -15,6 +15,7 @@
 // Auto-load environment variables from shared and local .env files
 import './env-loader.js';
 import debug from 'debug';
+import { extractFromHeaders, extractReferrer, getTokenSource, extractFromQuery, extractFromBody, extractToken, TOKEN_FIELDS } from './extractFromRequest.js';
 
 // Set up debug loggers with namespaces
 const log = debug('pollinations:auth');
@@ -22,133 +23,9 @@ const errorLog = debug('pollinations:error');
 const tokenLog = debug('pollinations:auth:token');
 const referrerLog = debug('pollinations:auth:referrer');
 
-// Token field configuration for DRY principle
-const TOKEN_FIELDS = {
-  query: ['token'],
-  header: ['authorization', 'x-pollinations-token'],
-  body: ['token' ]
-};
-
-/**
- * Helper function to extract value from query parameters
- * @param {string} url - Request URL
- * @param {string[]} fields - Array of field names to check
- * @returns {Object} { value, source } or { value: null, source: null }
- */
-function extractFromQuery(url, fields) {
-  const urlObj = new URL(url, 'http://x');
-  for (const field of fields) {
-    const value = urlObj.searchParams.get(field);
-    if (value) return { value, source: `query:${field}` };
-  }
-  return { value: null, source: null };
-}
-
-/**
- * Helper function to extract value from headers (supports both Cloudflare and Express styles)
- * @param {Object} headers - Request headers
- * @param {string[]} fields - Array of field names to check
- * @returns {Object} { value, source } or { value: null, source: null }
- */
-function extractFromHeaders(headers, fields) {
-  for (const field of fields) {
-    let value = null;
-    
-    if (headers.get) { // Cloudflare-style headers
-      value = headers.get(field);
-    } else if (headers) { // Express-style headers
-      value = headers[field] || headers[field.toLowerCase()];
-    }
-    
-    if (value) {
-      // Special handling for authorization header
-      if (field === 'authorization') {
-        value = value.replace(/^Bearer\s+/i, '');
-      }
-      return { value, source: `header:${field}` };
-    }
-  }
-  return { value: null, source: null };
-}
-
-/**
- * Helper function to extract value from request body
- * @param {Object} body - Request body
- * @param {string[]} fields - Array of field names to check
- * @returns {Object} { value, source } or { value: null, source: null }
- */
-function extractFromBody(body, fields) {
-  if (!body) return { value: null, source: null };
-  
-  for (const field of fields) {
-    const value = body[field];
-    if (value) return { value, source: `body:${field}` };
-  }
-  return { value: null, source: null };
-}
-
-/**
- * Extract authentication token from request
- * Supports multiple token sources (NO referrer fallback)
- * Compatible with OpenAI and other LLM API authentication patterns
- * @param {Request|Object} req - The request object
- * @returns {string|null} The token or null
- */
-export function extractToken(req) {
-  // Check URL query parameters
-  const queryResult = extractFromQuery(req.url, TOKEN_FIELDS.query);
-  if (queryResult.value) return queryResult.value;
-  
-  // Check headers for Bearer token, API key, and custom headers
-  const headerResult = extractFromHeaders(req.headers, TOKEN_FIELDS.header);
-  if (headerResult.value) return headerResult.value;
-  
-  // Check body for token field in POST requests (multiple field names)
-  if (req.method === 'POST' && req.body) {
-    const bodyResult = extractFromBody(req.body, TOKEN_FIELDS.body);
-    if (bodyResult.value) return bodyResult.value;
-  }
-  
-  return null;  // header/query/body only – no referrer here!
-}
-
-/**
- * Extract referrer from request headers and body
- * Used for frontend app identification, extended access, and analytics
- * @param {Request|Object} req - The request object (can be Cloudflare Request or Express req)
- * @returns {string|null} The referrer URL or null
- */
-export function extractReferrer(req) {
-  // First check URL query parameters (highest priority)
-  const url = req.url;
-  if (url) {
-    const urlObj = new URL(url, 'http://x'); // Use dummy base for relative URLs
-    const queryReferrer = urlObj.searchParams.get('referrer') || 
-                         urlObj.searchParams.get('referer'); // Support both spellings
-    if (queryReferrer) return queryReferrer;
-  }
-  
-  // Then check body for referrer field (second priority) - not just for POST
-  if (req.body?.referrer) return String(req.body.referrer);
-  if (req.body?.referer) return String(req.body.referer);
-  
-  // Finally check headers (lowest priority)
-  // Handle Cloudflare Workers Request
-  if (req.headers && typeof req.headers.get === 'function') {
-    const headerReferrer = req.headers.get('referer') || 
-                          req.headers.get('referrer') || // Support both spellings
-                          req.headers.get('origin');
-    if (headerReferrer) return headerReferrer;
-  } else if (req.headers && typeof req.headers === 'object') {
-    // Handle Express/Node.js request
-    const headerReferrer = req.headers.referer || 
-                          req.headers.referrer || // Support both spellings
-                          req.headers.origin;
-    if (headerReferrer) return headerReferrer;
-  }
-  
-  return null;
-}
+// Extract authentication token from request
+// This function has been moved to extractFromRequest.js
+// and is now imported at the top of this file
 
 /**
  * Validate token against allowed tokens
@@ -170,101 +47,55 @@ export function isValidToken(token, validTokens) {
 }
 
 /**
- * Determine the source of the token (header, query param, body)
- * @param {Request|Object} req - The request object
- * @returns {string} The source of the token
+ * Check if a referrer domain is registered by any user in the auth.pollinations.ai database
+ * @param {string} referrer - The referrer URL to check
+ * @returns {Promise<{userId: string, username: string, tier: string}|null>} User info if domain is registered, null otherwise
  */
-export function getTokenSource(req) {
-  const token = extractToken(req);
-  if (!token) return 'unknown';
-
-  const queryResult = extractFromQuery(req.url, TOKEN_FIELDS.query);
-  if (queryResult.value === token) return queryResult.source;
-
-  const headerResult = extractFromHeaders(req.headers, TOKEN_FIELDS.header);
-  if (headerResult.value === token) return headerResult.source;
-
-  const bodyResult = extractFromBody(req.body, TOKEN_FIELDS.body);
-  if (bodyResult.value === token) return bodyResult.source;
-
-  return 'unknown';
-}
-
-/**
- * Get client IP address from request
- * @param {Request|Object} req - The request object
- * @returns {string} The client IP or 'unknown'
- */
-export function getClientIp(req) {
-  // Handle Cloudflare Workers Request
-  if (req.headers && typeof req.headers.get === 'function') {
-    return req.headers.get('cf-connecting-ip') ||
-           req.headers.get('x-real-ip') ||
-           req.headers.get('x-forwarded-for')?.split(',')[0].trim() ||
-           'unknown';
-  }
+export async function checkReferrerInDb(referrer) {
+  if (!referrer) return null;
   
-  // Handle Express/Node.js request
-  if (req.headers && typeof req.headers === 'object') {
-    return req.headers['cf-connecting-ip'] ||
-           req.headers['x-real-ip'] ||
-           (req.headers['x-forwarded-for'] || '').split(',')[0].trim() ||
-           req.connection?.remoteAddress ||
-           'unknown';
-  }
-  
-  return 'unknown';
-}
-
-/**
- * Get IP address from request (with privacy masking)
- * @param {Object} req - The request object
- * @returns {string} The IP address (truncated for privacy)
- */
-export function getIp(req) {
-  // Prioritize standard proxy headers and add cloudflare-specific headers
-  const ip = req.headers["x-bb-ip"] || 
-             req.headers["x-nf-client-connection-ip"] || 
-             req.headers["x-real-ip"] || 
-             req.headers['x-forwarded-for'] || 
-             req.headers['cf-connecting-ip'] ||
-             (req.socket ? req.socket.remoteAddress : null);
-  
-  if (!ip) return null;
-  
-  // Handle x-forwarded-for which can contain multiple IPs (client, proxy1, proxy2, ...)
-  // The client IP is typically the first one in the list
-  const cleanIp = ip.split(',')[0].trim();
-  
-  // Check if IPv4 or IPv6
-  if (cleanIp.includes(':')) {
-      // IPv6 address - take first 4 segments (64 bits) which typically represent the network prefix
-      // Handle special IPv6 formats like ::1 or 2001::
-      const segments = cleanIp.split(':');
-      let normalizedSegments = [];
-      
-      // Handle :: notation (compressed zeros)
-      if (cleanIp.includes('::')) {
-          const parts = cleanIp.split('::');
-          const leftPart = parts[0] ? parts[0].split(':') : [];
-          const rightPart = parts[1] ? parts[1].split(':') : [];
-          
-          // Calculate how many zero segments are represented by ::
-          const missingSegments = 8 - leftPart.length - rightPart.length;
-          
-          normalizedSegments = [
-              ...leftPart,
-              ...Array(missingSegments).fill('0'),
-              ...rightPart
-          ];
-      } else {
-          normalizedSegments = segments;
+  try {
+    // Extract domain from referrer URL if it's a full URL
+    let domain = referrer;
+    if (referrer.startsWith('http://') || referrer.startsWith('https://')) {
+      try {
+        const urlObj = new URL(referrer);
+        domain = urlObj.hostname;
+      } catch (error) {
+        // If parsing fails, use the raw referrer value
+        referrerLog('Failed to parse referrer as URL, using raw value:', error);
       }
-      
-      return normalizedSegments.slice(0, 4).join(':');
-  } else {
-      // IPv4 address - take first 3 segments
-      return cleanIp.split('.').slice(0, 3).join('.');
+    }
+    
+    referrerLog(`Checking if domain ${domain} is registered in auth.pollinations.ai database`);
+    
+    // Query the auth.pollinations.ai API to check if the domain is registered by any user
+    const apiUrl = `https://auth.pollinations.ai/api/validate-referrer?referrer=${encodeURIComponent(domain)}`;
+    referrerLog(`Calling API: ${apiUrl}`);
+    
+    const response = await fetch(apiUrl);
+    if (!response.ok) {
+      referrerLog(`API returned error status: ${response.status}`);
+      return null;
+    }
+    
+    const data = await response.json();
+    referrerLog('API response:', data);
+    
+    if (data && data.valid && data.user_id) {
+      referrerLog(`✅ Domain ${domain} is registered by user ${data.user_id} (${data.username}) with tier ${data.tier}`);
+      return {
+        userId: data.user_id,
+        username: data.username,
+        tier: data.tier
+      };
+    } else {
+      referrerLog(`❌ Domain ${domain} is not registered in the database`);
+      return null;
+    }
+  } catch (error) {
+    referrerLog('Error checking referrer in database:', error);
+    return null;
   }
 }
 
@@ -493,23 +324,6 @@ export async function shouldBypassQueue(req, { legacyTokens, allowlist }) {
       referrerLog('No legacy token found in referrer');
     }
   
-    // 3.5️⃣ Special check for catgpt referrer
-    if (refStr.toLowerCase().includes('pollinations.github.io')) {
-      referrerLog('✅ CatGPT referrer detected: %s', refStr);
-      debugInfo.authResult = 'CATGPT_REFERRER';
-      debugInfo.catgptMatch = true;
-      log('Authentication succeeded: CATGPT_REFERRER');
-      return { 
-        bypass: true, // Kept for backward compatibility
-        authenticated: true, 
-        tokenAuth: false, 
-        referrerAuth: true,
-        reason: 'CATGPT_REFERRER', 
-        userId: null, 
-        debugInfo 
-      };
-    }
-  
     // 4️⃣ Check allow-listed domain
     referrerLog('Checking referrer against %d allowlisted domains', debugInfo.allowlistCount);
     const allowlistMatch = allowlist.some(d => refStr.includes(d));
@@ -519,7 +333,7 @@ export async function shouldBypassQueue(req, { legacyTokens, allowlist }) {
       debugInfo.allowlistMatch = true;
       log('Authentication succeeded: ALLOWLIST');
       return { 
-        bypass: true, // Kept for backward compatibility
+        bypass: true, // Enabling bypass for allowlisted domains
         authenticated: true, 
         tokenAuth: false, 
         referrerAuth: true,
@@ -529,6 +343,31 @@ export async function shouldBypassQueue(req, { legacyTokens, allowlist }) {
       };
     } else {
       referrerLog('Referrer does not match any allowlisted domain');
+    }
+    
+    // 5️⃣ Check for domain registered in auth database
+    referrerLog('Checking if referrer is registered in auth database');
+    const dbReferrerResult = await checkReferrerInDb(ref);
+    if (dbReferrerResult && dbReferrerResult.userId) {
+      referrerLog('✅ Registered domain: %s for user %s (tier: %s)', ref, dbReferrerResult.userId, dbReferrerResult.tier);
+      debugInfo.authResult = 'DB_REFERRER';
+      debugInfo.dbReferrerMatch = true;
+      debugInfo.userId = dbReferrerResult.userId;
+      debugInfo.username = dbReferrerResult.username;
+      debugInfo.tier = dbReferrerResult.tier;
+      log('Authentication succeeded: DB_REFERRER for user %s (tier: %s)', dbReferrerResult.userId, dbReferrerResult.tier);
+      return { 
+        bypass: true, // Enable bypass for DB-registered referrers
+        authenticated: true, 
+        tokenAuth: false, 
+        referrerAuth: true,
+        reason: 'DB_REFERRER', 
+        userId: dbReferrerResult.userId,
+        tier: dbReferrerResult.tier, 
+        debugInfo 
+      };
+    } else {
+      referrerLog('Referrer is not registered in auth database');
     }
   }
